remove _vectorized
