+++ geopandas/testing.py
@@ -7,7 +7,6 @@ import pandas as pd
 
 from geopandas import GeoDataFrame, GeoSeries
 from geopandas.array import GeometryDtype
-from geopandas import _vectorized
 
 
 def _isna(this):
@@ -186,8 +185,8 @@ def assert_geoseries_equal(
         )
 
     if normalize:
-        left = GeoSeries(_vectorized.normalize(left.array._data))
-        right = GeoSeries(_vectorized.normalize(right.array._data))
+        left = GeoSeries(left.array.normalize())
+        right = GeoSeries(right.array.normalize())
 
     if not check_crs:
         with warnings.catch_warnings():
